Add files via upload
The supply chain is one of the domains in that data analysis could be useful.
Data in the supply chain is mostly in a form of transaction data that each row represents 1 transaction of 1 item and 1 customer.
In this dashboard  I illustrated how I analyze the supply chain data. Hope that it will give some ideas to anyone who works on this type of data
For this supply chain dataset, I used Power BI as a visualization tool since it is easy to filter by a particular category of product or customer.